new endpoints
Some new endpoints has been added to the api in order to get the actually forum started.

+++ api/account/destroy.php
@@ -5,9 +5,9 @@ error_reporting(E_ALL);
 ini_set('display_errors', 1);
 
 header('Access-Control-Allow-Origin: *');
-header('Content-Type: application/json');
+//header('Content-Type: application/json');
 header('Access-Control-Allow-Headers: *');
-//header('Content-Type: text/html');
+header('Content-Type: text/html');
 header('Access-Control-Allow-Methods: *');
 
 // Including required files
@@ -24,9 +24,67 @@ $data = json_decode(file_get_contents('php://input'));
 
 if(isset($data)) {
 
-    if($account->destroy($data->account_id)) {
-        echo json_encode(['message' => 'Role has been deleted successfully']);
-    } else {
-        echo json_encode(['message' => 'Role could not be deleted']);
+    //check if account exists in the system
+    $account_data = $account->accountExists($data->account_id);
+    while($row = $account_data->fetch(PDO::FETCH_OBJ)) {
+        $role_id = $row->role_id;
+        //echo 'Role ID: ' . $role_id;
+        $role_data = $account->checkRole($role_id);
+        while($row1 = $role_data->fetch(PDO::FETCH_OBJ)) {
+            $rolename = $row1->role_name;
+            //echo 'Role: ' .$rolename;
+            if($rolename != 'Owner') {
+                if($account->destroy($data->account_id)) {
+                    if($account->clearLoginHistory($data->account_id)) {
+                        $logmsg = "Login history has been cleared out!";
+                    } else {
+                        $logmsg = "Something went wrong";
+                    }
+                    if($account->deleteProfile($data->account_id)) {
+                        $promsg = "Connected profile is deleted";
+                    } else {
+                        $promsg = "something went wrong";
+                    }
+                    echo json_encode(['message' => 'Account has been deleted successfully', 'logmsg' => $logmsg, 'promsg' => $promsg]);
+                } else {
+                    echo json_encode(['message' => 'Account could not be deleted']);
+                }
+
+            } else {
+                echo json_encode(array('message' => 'An owner can not be deleted'));
+            }
+        }
     }
+} else if(isset($_GET['account_id'])) {
+    $account_data = $account->accountExists($_GET['account_id']);
+    while($row = $account_data->fetch(PDO::FETCH_OBJ)) {
+        $role_id = $row->role_id;
+        //echo 'Role ID: ' . $role_id;
+        $role_data = $account->checkRole($role_id);
+        while($row1 = $role_data->fetch(PDO::FETCH_OBJ)) {
+            $rolename = $row1->role_name;
+            //echo 'Role: ' .$rolename;
+            if($rolename != 'Owner') {
+                if($account->destroy($_GET['account_id'])) {
+                    if($account->clearLoginHistory($_GET['account_id'])) {
+                        $logmsg = "Login history has been cleared out!";
+                    } else {
+                        $logmsg = "Something went wrong";
+                    }
+                    if($account->deleteProfile($_GET['account_id'])) {
+                        $promsg = "Connected profile is deleted";
+                    } else {
+                        $promsg = "something went wrong";
+                    }
+                    echo json_encode(['message' => 'Account has been deleted successfully', 'logmsg' => $logmsg, 'promsg' => $promsg]);
+                } else {
+                    echo json_encode(['message' => 'Account could not be deleted']);
+                }
+            } else {
+                echo json_encode(array('message' => 'An owner can not be deleted'));
+            }
+        }
+    }
+
+
 }
